feat: switch from trade to webserver command in docker-compose
Changed the default command in docker-compose.yml from 'trade' to 'webserver'
to enable web interface functionality. Removed unnecessary database URL and
strategy parameters that are not required for the webserver command, keeping
only the essential logfile and config parameters. This allows users to access
FreqTrade through the web UI when running the container.